feat: add author format option to fzf-git-branches
This commit introduces a new option to the fzf-git-branches script,
allowing the user to specify the format for the author string. This
enhancement provides more flexibility to the user in terms of how they
want to view the author information in the output.

The new option can be used with the `-u` or `--author-format` flags,
followed by the desired format. The default format is "committername".