feat: add function to transform git format strings
A new function, `__fgb_transform_git_format_string`, has been added.
This function takes a Git format string as input and transforms it if
necessary.

The transformation is applied to `c_date_format` and `c_author_format`
in both `__fgb_branch` and `__fgb_worktree` functions. This change aims
to improve the flexibility of format string handling in the script.