feat: Improve parsing of branch, author, date
This commit refactors the way we parse the branch name, author name, and
author date from the git output. Instead of using a single read command,
we now use string manipulation to extract the required information. This
change slightly improves performance according to the tests done. It
also improves the robustness of the parsing process and makes the code
easier to understand and maintain.