feat: use variable expansion instead of sed for fzf preview
This commit changes the way the branch name is passed to the preview
command. Instead of using a sed command to remove surrounding brackets
form the branch name, it now uses a variable (FGB_BRANCH) to hold the
branch name and shell variable expansion. This makes the code more
readable and maintainable.