feat: Add worktree for new branch
This commit introduces a new feature that allows the creation of a
worktree for a new branch. This is achieved by adding a new function
`__fgb_git_worktree_for_new_branch`. This function is then integrated
into the existing workflow by adding a new case in the switch statement
for the key `$c_new_branch_key`. This key is also added to the expected
keys for the `fzf_cmd` in the `__fgb_worktree_add` and
`__fgb_worktree_manage` functions. The variable `c_new_branch` is
introduced to store the name of the new branch.