feat: Add stash restore to worktree creation
This commit introduces a new feature to the script. When creating a new
worktree for a new branch, the script now checks if there are any
uncommitted changes in the current worktree. If there are, it offers to
stash them and apply the stash in the new worktree. This feature is
useful when you want to switch branches but have uncommitted changes
that you don't want to commit yet.